+++ CHANGELOG.rst
@@ -2,11 +2,23 @@
 Changelog
 =========
 
+Version 22.4
+============
+
+* Python 3.10 support is deprecated and will be removed in the future.
+  `#173 <https://github.com/iqm-finland/iqm-client/pull/173>`_
+
+Version 22.3
+============
+
+* Add the native ``reset`` operation. `#170 <https://github.com/iqm-finland/iqm-client/pull/170>`_
+
 Version 22.2
 ============
 
 * Improve the documentation of the ``delay`` operation.
 * Improve the performance of :func:`.transpile_insert_moves`.
+  `#168 <https://github.com/iqm-finland/iqm-client/pull/168>`_
 
 Version 22.1
 =============